pyproject.toml
[project]
name = "colorbyte-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.129.0",
    "gradio-client>=2.0.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.0",
    "python-multipart>=0.0.22",
    "replicate>=1.0.7",
    "uvicorn>=0.40.0",
]
